HARP-6096: Add fast conversions between the WEB Mercator projections.
Also, this change places the "WEB Mercator"-projection variants
currently supported by @here/harp-geoutils in the same file. This is
needed to avoid possible circular dependencies (e.g reprojectPoint()
needs to access both webMercatorProjection and mercatorProjection).

Placing both variants in the same file has the nice side-effect to
remove the need to export the class MercatorProjection.
