+++ @here/harp-map-controls/test/MapControlsTest.ts
@@ -7,7 +7,7 @@
 // tslint:disable:only-arrow-functions
 //    Mocha discourages using arrow functions, see https://mochajs.org/#arrow-functions
 
-import { MercatorProjection } from "@here/harp-geoutils";
+import { mercatorProjection } from "@here/harp-geoutils";
 import { MapView } from "@here/harp-mapview";
 import { expect } from "chai";
 import * as sinon from "sinon";
@@ -66,7 +66,7 @@ describe("MapControls", function() {
             sandbox.stub(mapView, "maxZoomLevel").get(() => maxZoom);
             sandbox.stub(mapView, "minZoomLevel").get(() => minZoom);
             sandbox.stub(mapView, "minCameraHeight").get(() => minCameraHeight);
-            sandbox.stub(mapView, "projection").get(() => MercatorProjection);
+            sandbox.stub(mapView, "projection").get(() => mercatorProjection);
 
             mapControls = new MapControls(mapView);
         });
